+++ src/_installation/02-create-and-fill-users-table.sql
@@ -5,6 +5,7 @@ CREATE TABLE IF NOT EXISTS `formsdb`.`users` (
   `user_email` varchar(64) COLLATE utf8_unicode_ci NOT NULL COMMENT 'user''s email, unique',
   `user_group` varchar(64) COLLATE utf8_unicode_ci NOT NULL COMMENT 'group name user belongs to',
   `form_manager` boolean COLLATE utf8_unicode_ci NOT NULL default 0 COMMENT 'user can manage forms',
+  `user_perms` TINYINT NOT NULL DEFAULT '3' COMMENT 'smaller=higher perms',
   PRIMARY KEY (`user_id`),
   UNIQUE KEY `user_name` (`user_name`),
   UNIQUE KEY `user_email` (`user_email`)