+++ src/includes/pageformsbuild.php
@@ -237,6 +237,7 @@
     </div>
 
 </div>
+hello
 <!-- /#page-wrapper -->
 
         <script>